+++ include/pajlada/settings/signalargs.hpp
@@ -19,6 +19,17 @@ struct SignalArgs {
 
     bool writeToFile{true};
     bool compareBeforeSet{false};
+
+private:
+    /// If set to true, the SettingManager instance will attempt to
+    /// only remove the setting from the RapidJSON document instead of
+    /// setting the value to its default value.
+    bool resetToDefault{false};
+
+    friend class SettingManager;
+
+    template <typename Type>
+    friend class Setting;
 };
 
 }  // namespace pajlada::Settings