+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Breaking: Saving a setting that was previously set and is then reset to its default value with `resetToDefaultValue` will now omit that key if possible, instead of saving the Setting's default value in the JSON file. (#175)
+
 ## v0.5.0
 
 - Breaking: In CMake, PajladaSerialize and PajladaSignals are now private so you need to include them yourself. (#164)